renamed the plugin, also made a structured tips.
tips would be a matrix/2d list now, but instead it would behave like a
linear list.
to add something to the tips, we would just insert a table containing
the new tips in the `tips` table. but when returning a random tip, the
ReturnRandomTip() calculates the inner length of the list, and then
calculates the x and y of the matrix like it was a graph. makes
visualizing the matrix more easy.